### 2025-07-07 #8549 @editorialbot
[REVIEW]: jaxhps: An elliptic PDE solver built with machine learning in mind
**Submitting author:** @meliao (<a href="http://orcid.org/0000-0003-0737-3718">Owen Melia</a>)
**Repository:** https://github.com/meliao/jaxhps
**Branch with paper.md** (empty if default branch): JOSS_submission
**Version:** v0.1
