+++ tests/test_local_solve/test_local_solve_uniform_2D_DtN.py
@@ -6,9 +6,14 @@ from jaxhps.local_solve._uniform_2D_DtN import (
     local_solve_stage_uniform_2D_DtN,
     _gather_coeffs_2D,
     assemble_diff_operator,
+    assemble_diff_operator_rectangular,
     get_DtN,
+    get_DtN_rectangular,
 )
-from jaxhps._precompute_operators_2D import precompute_diff_operators_2D
+from jaxhps._precompute_operators_2D import (
+    precompute_diff_operators_2D,
+    precompute_rectangular_diff_operators_2D,
+)
 from jaxhps._discretization_tree import DiscretizationNode2D
 from jaxhps._domain import Domain
 from jaxhps._pdeproblem import PDEProblem
@@ -72,11 +77,28 @@ class Test_assemble_diff_operator:
         jax.clear_caches()
 
 
-class Test__local_solve_stage_2D:
+class Test_assemble_diff_operator_rectangular:
+    def test_0(self) -> None:
+        p = 8
+        half_side_len = 0.25
+
+        d_x, d_y, d_xx, d_yy, d_xy, b = (
+            precompute_rectangular_diff_operators_2D(p, half_side_len)
+        )
+
+        stacked_diff_operators = jnp.stack([d_xx, d_xy, d_yy, d_x, d_y, b])
+        coeffs_arr = jnp.ones((3, p**2))
+        which_coeffs = jnp.array([True, True, True, False, False, False])
+        out = assemble_diff_operator_rectangular(
+            coeffs_arr, which_coeffs, stacked_diff_operators
+        )
+        assert out.shape == ((p - 2) ** 2, p**2)
+        jax.clear_caches()
+
+
+class Test_local_solve_stage_uniform_2D_DtN:
     def test_0(self, caplog) -> None:
-        """Tests the solve_stage function returns without error and returns the correct shape
-        when boundary data are not passed to the function.
-        """
+        """Tests the solve_stage function returns without error and returns the correct shape."""
         caplog.set_level(logging.DEBUG)
         p = 8
         q = 6
@@ -114,6 +136,48 @@ class Test__local_solve_stage_2D:
         assert h_arr.shape == (n_leaves, 4 * q)
         jax.clear_caches()
 
+    def test_1(self, caplog) -> None:
+        """Tests the solve_stage function returns without error and returns the correct shape
+        in the rectangular spectral collocation case.
+        """
+        caplog.set_level(logging.DEBUG)
+        p = 8
+        q = 6
+        l = 2
+        n_leaves = 4**l
+
+        root = DiscretizationNode2D(xmin=0.0, xmax=1.0, ymin=0.0, ymax=1.0)
+        domain = Domain(p=p, q=q, root=root, L=l)
+
+        # This could be np.random.normal(size=(n_leaves, p**2))
+        d_xx_coeffs = jnp.array(np.random.normal(size=(n_leaves, p**2)))
+        d_yy_coeffs = jnp.array(np.random.normal(size=(n_leaves, p**2)))
+        source_term = jnp.array(np.random.normal(size=(n_leaves, p**2)))
+
+        t = PDEProblem(
+            domain=domain,
+            source=source_term,
+            D_xx_coefficients=d_xx_coeffs,
+            D_yy_coefficients=d_yy_coeffs,
+            use_rectangular_spectral_collocation=True,
+        )
+
+        print("test_1: n_leaves = ", n_leaves)
+        print("test_1: n_gauss_bdry = ", 4 * q)
+
+        (
+            Y_arr,
+            T_arr,
+            v_arr,
+            h_arr,
[…]
