### 2025-07-15 84a0a84
Didn't find bug but did add more tests
