### 2025-07-15 1711bce
2D uniform DtN works with rectangular spectral collocation
