+++ tests/test_accuracy/test_local_solve_stage_accuracy.py
@@ -11,10 +11,8 @@ from jaxhps.local_solve._nosource_uniform_2D_ItI import (
     nosource_local_solve_stage_uniform_2D_ItI,
 )
 from jaxhps.local_solve._uniform_2D_ItI import local_solve_stage_uniform_2D_ItI
-import matplotlib.pyplot as plt
 
 
-# from jaxhps._utils import plot_soln_from_cheby_nodes
 from .cases import (
     XMIN,
     XMAX,
@@ -195,15 +193,15 @@ def check_leaf_accuracy_DtN(
     )
 
     # Plot the expected and computed outgoing data
-    plt.plot(computed_outgoing_data, label="Computed Outgoing Data")
-    plt.plot(expected_outgoing_data, label="Expected Outgoing Data")
-    plt.legend()
-    plt.show()
-    plt.clf()
+    # plt.plot(computed_outgoing_data, label="Computed Outgoing Data")
+    # plt.plot(expected_outgoing_data, label="Expected Outgoing Data")
+    # plt.legend()
+    # plt.show()
+    # plt.clf()
 
-    # assert jnp.allclose(
-    #     computed_outgoing_data, expected_outgoing_data, atol=ATOL, rtol=RTOL
-    # )
+    assert jnp.allclose(
+        computed_outgoing_data, expected_outgoing_data, atol=ATOL, rtol=RTOL
+    )
 
     ##############################################################
     # Check the accuracy of the homogeneous solution
@@ -220,6 +218,16 @@ def check_leaf_accuracy_DtN(
         "check_leaf_accuracy_DtN: expected_homogeneous_soln shape: %s",
         expected_homogeneous_soln.shape,
     )
+
+    # plot it
+    # plot_soln_from_cheby_nodes(
+    #     cheby_nodes=domain.interior_points.reshape(-1, 2),
+    #     corners=None,
+    #     expected_soln=expected_homogeneous_soln.real.flatten(),
+    #     computed_soln=computed_homogeneous_soln.real.flatten(),
+    #     t="Homogeneous Solution",
+    # )
+
     assert jnp.allclose(
         computed_homogeneous_soln,
         expected_homogeneous_soln,
@@ -244,6 +252,15 @@ def check_leaf_accuracy_DtN(
     assert jnp.allclose(
         computed_part_soln, expected_part_soln, atol=ATOL, rtol=RTOL
     )
+
+    # Plot it
+    # plot_soln_from_cheby_nodes(
+    #     cheby_nodes=domain.interior_points.reshape(-1, 2),
+    #     corners=None,
+    #     expected_soln=expected_part_soln.real.flatten(),
+    #     computed_soln=computed_part_soln.real.flatten(),
+    #     t="Particular Solution",
+    # )
 
     logging.debug("check_leaf_accuracy_DtN: passed part soln check")
     ##############################################################
@@ -272,11 +289,11 @@ def check_leaf_accuracy_DtN(
     )
 
     # Plot the expected and computed outgoing data
-    plt.plot(computed_outgoing_part_data.real, label="Computed Outgoing Data")
-    plt.plot(expected_outgoing_part_data.real, label="Expected Outgoing Data")
-    plt.legend()
-    plt.show()
-    plt.clf()
+    # plt.plot(computed_outgoing_part_data.real, label="Computed Outgoing Data")
+    # plt.plot(expected_outgoing_part_data.real, label="Expected Outgoing Data")
+    # plt.legend()
+    # plt.show()
+    # plt.clf()
     assert jnp.allclose(
         computed_outgoing_part_data,
         expected_outgoing_part_data,
