### 2025-07-15 c9ca942
Adding incomplete reverse mode autodiff example
