+++ src/jaxhps/_precompute_operators_2D.py
@@ -45,15 +45,17 @@ def precompute_diff_operators_2D(
     )
 
 
-def rectangular_interp_operator(p: int) -> jax.Array:
+def rectangular_interp_operator(p: int) -> Tuple[jax.Array, jax.Array]:
     """
     Returns an interpolation operator that maps from a grid of p^2 2nd-kind Chebyshev points to a grid of (p-2)^2 1st-kind Chebyshev points.
+
+    Also returns an interpolation matrix that maps from the 1st-kind Chebyshev points to the 2nd-kind Chebyshev points.
 
     Args:
         p (int): Number of points per dimension on the 2nd-kind Chebyshev grid.
 
     Returns:
-        jax.Array: Has shape ((p - 2) ** 2, p ** 2).
+        jax.Array: 2nd->1st has shape ((p - 2) ** 2, p ** 2). 1st->2nd has shape (p ** 2, (p - 2) ** 2).
     """
 
     # Precompute a 2D barycentric Lagrange interpolation matrix
@@ -68,7 +70,14 @@ def rectangular_interp_operator(p: int) -> jax.Array:
     # Rearrange the cols of the interpolation matrix to match the expected grid ordering
     rearrange_indices = rearrange_indices_ext_int(p)
     B = B[:, rearrange_indices]
-    return B
+
+    # D is the interpolation matrix from the 1st-kind Chebyshev points to the 2nd-kind Chebyshev points.
+    D = barycentric_lagrange_interpolation_matrix_2D(
+        to_x, to_y, from_x, from_y
+    )
+    # Rearrange the rows of the interpolation matrix to match the expected grid ordering
+    D = D[rearrange_indices, :]
+    return B, D
 
 
 def precompute_rectangular_diff_operators_2D(
@@ -93,16 +102,16 @@ def precompute_rectangular_diff_operators_2D(
 
     D_x, D_y, D_xx, D_yy, D_xy = precompute_diff_operators_2D(p, half_side_len)
 
-    B = rectangular_interp_operator(p)
+    B, D = rectangular_interp_operator(p)
 
     # Apply the barycentric Lagrange interpolation matrix to the differentiation operators
-    D_x = B @ D_x
-    D_y = B @ D_y
-    D_xx = B @ D_xx
-    D_yy = B @ D_yy
-    D_xy = B @ D_xy
+    D_x = interpolate_cheby_to_cheby_first_kind(B, D, D_x)
+    D_y = interpolate_cheby_to_cheby_first_kind(B, D, D_y)
+    D_xx = interpolate_cheby_to_cheby_first_kind(B, D, D_xx)
+    D_yy = interpolate_cheby_to_cheby_first_kind(B, D, D_yy)
+    D_xy = interpolate_cheby_to_cheby_first_kind(B, D, D_xy)
 
-    return (D_x, D_y, D_xx, D_yy, D_xy, B)
+    return (D_x, D_y, D_xx, D_yy, D_xy, B, D)
 
 
 # Unclear whether this is a win to jit this one.
@@ -438,3 +447,56 @@ def indexing_for_refinement_operator(p: int) -> jnp.array:
         ]
     )
     return row_idxes, col_idxes
+
+
+@jax.custom_vjp
+def interpolate_cheby_to_cheby_first_kind(
+    B: jax.Array, D: jax.Array, v: jax.Array
+) -> jax.Array:
+    """
+    Interpolates a function from the Chebyshev grid of the second kind to the Chebyshev grid of the first kind.
+
+    Really all it does is B @ v. Separated it out to allow for a custom autodiff rule.
+
+    Args:
[…]
