### 2025-07-28 0304dc2
Added untested functionality for computing the incoming particular solution data during the upward pass for 2D ItI problems
