+++ src/jaxhps/up_pass/_uniform_2D_ItI.py
@@ -10,7 +10,7 @@ def up_pass_uniform_2D_ItI(
     pde_problem: PDEProblem,
     device: jax.Device = jax.devices()[0],
     host_device: jax.Device = jax.devices("cpu")[0],
-    return_h_last: bool = False,
+    return_bdry_imp_data: bool = False,
 ) -> Tuple[jax.Array, List[jax.Array], jax.Array]:
     """
     This function performs the upward pass for 2D ItI problems. It recomputes the local solve stage to get
@@ -32,8 +32,8 @@ def up_pass_uniform_2D_ItI(
     host_device : jax.Device, optional
         Where to place the output. Defaults to ``jax.devices("cpu")[0]``.
 
-    return_h_last : bool, optional
-        If True, return the last h vector, which gives the outgoing impedance data for the particular solution evaluated at the domain boundary.
+    return_bdry_imp_data : bool, optional
+        If True, return outgoing and incoming impedance data for the particular solution evaluated at the domain boundary.
 
     Returns
     -------
@@ -44,7 +44,10 @@ def up_pass_uniform_2D_ItI(
         List of pre-computed g_tilde matrices for each level of the quadtree.
 
     h_last : jax.Array
-        Outgoing impedance data for the particular solution evaluated at the domain boundary. Has shape (nbdry, nsrc). Is only returned if ``return_h_last=True``.
+        Outgoing impedance data for the particular solution evaluated at the domain boundary. Has shape (nbdry, nsrc). Is only returned if ``return_bdry_imp_data=True``.
+
+    g_last : jax.Array
+        Incoming impedance data for the particular solution evaluated at the domain boundary. Has shape (nbdry, nsrc). Is only returned if ``return_bdry_imp_data=True``.
     """
 
     bool_multi_source = source.ndim == 3
@@ -86,12 +89,8 @@ def up_pass_uniform_2D_ItI(
     # Get leaf-level h_in array, which is an einsum between the particular soln and
     # the QH array stored in the PDEProblem.
     h_in = jnp.einsum("ij,kjl->kil", pde_problem.QH, v)
+    g_in = jnp.einsum("ij,kjl->kil", pde_problem.QG, v)
 
-    # Y, T, v, h_in = local_solve_stage_uniform_2D_ItI(
-    #     pde_problem=pde_problem,
-    #     device=device,
-    #     host_device=host_device,
-    # )
     logging.debug(
         "up_pass_uniform_2D_ItI: after local solve, h_in shape = %s",
         h_in.shape,
@@ -104,10 +103,13 @@ def up_pass_uniform_2D_ItI(
 
         nnodes, nbdry, nsrc = h_in.shape
         h_in = h_in.reshape(nnodes // 4, 4, nbdry, nsrc)
+        g_in = g_in.reshape(nnodes // 4, 4, nbdry, nsrc)
         D_inv = D_inv_lst[i]
         BD_inv = BD_inv_lst[i]
 
-        h_in, g_tilde = vmapped_assemble_boundary_data(h_in, D_inv, BD_inv)
+        h_in, g_in, g_tilde = vmapped_assemble_boundary_data(
+            h_in, g_in, D_inv, BD_inv
+        )
         g_tilde_lst.append(g_tilde)
 
     logging.debug(
@@ -123,6 +125,7 @@ def up_pass_uniform_2D_ItI(
             jnp.squeeze(g_tilde, axis=-1) for g_tilde in g_tilde_lst
         ]
[…]
