### 2025-07-29 e40af3d
More work on test cases for 2D ItI problems
