+++ tests/test_accuracy/test_ItI_up_pass_accuracy.py
@@ -3,13 +3,16 @@ from typing import Dict
 import jax.numpy as jnp
 import jax
 import matplotlib.pyplot as plt
-
+import pytest
 
 from jaxhps._discretization_tree import DiscretizationNode2D
 from jaxhps._domain import Domain
 from jaxhps._pdeproblem import PDEProblem
 from jaxhps._build_solver import build_solver
 from jaxhps.up_pass._uniform_2D_ItI import up_pass_uniform_2D_ItI
+from jaxhps.local_solve._nosource_uniform_2D_ItI import (
+    nosource_local_solve_stage_uniform_2D_ItI,
+)
 # from jaxhps._utils import plot_soln_from_cheby_nodes
 
 from .cases import (
@@ -93,8 +96,12 @@ def check_up_pass_accuracy_2D_ItI_uniform_Helmholtz_like(
     # Plot outgoing_imp_data against h_last
     plt.plot(outgoing_imp_data.real, label="Outgoing Impedance Data.real")
     plt.plot(outgoing_imp_data.imag, label="Outgoing Impedance Data.imag")
+    # plt.plot(incoming_imp_data.real, label="Incoming Impedance Data.real")
+    # plt.plot(incoming_imp_data.imag, label="Incoming Impedance Data.imag")
     plt.plot(h_last.real, label="h_last.real")
     plt.plot(h_last.imag, label="h_last.imag")
+    # plt.plot(g_last.real, label="g_last.real")
+    # plt.plot(g_last.imag, label="g_last.imag")
     plt.legend()
     plt.show()
 
@@ -106,11 +113,102 @@ def check_up_pass_accuracy_2D_ItI_uniform_Helmholtz_like(
 
 
 class Test_accuracy_up_pass_2D_ItI_uniform:
+    @pytest.mark.skip
     def test_0(self, caplog) -> None:
-        """Complex coefficients."""
+        """Testing whether the computed outgoing and incoming particular soln impedance data matches against expectatiosn after 1 merge level."""
         caplog.set_level(logging.DEBUG)
 
         check_up_pass_accuracy_2D_ItI_uniform_Helmholtz_like(
             DOMAIN_ITI_NONPOLY, TEST_CASE_HELMHOLTZ_ITI_COMPLEX_COEFFS
         )
+        jax.clear_caches()
+
+    def test_1(self, caplog) -> None:
+        """Testing whether the computed outgoing and incoming particular soln impedance data matches against expectatiosn after the local solve."""
+        caplog.set_level(logging.DEBUG)
+
+        domain = Domain(p=P, q=Q, root=ROOT_ITI, L=0)
+
+        test_case = TEST_CASE_HELMHOLTZ_ITI_COMPLEX_COEFFS
+        d_xx_coeffs = test_case[K_XX_COEFF](domain.interior_points)
+        d_yy_coeffs = test_case[K_YY_COEFF](domain.interior_points)
+        i_coeffs = test_case[K_I_COEFF](domain.interior_points)
+
+        pde_problem = PDEProblem(
+            domain=domain,
+            D_xx_coefficients=d_xx_coeffs,
+            D_yy_coefficients=d_yy_coeffs,
+            I_coefficients=i_coeffs,
+            use_ItI=True,
+            eta=ETA,
+        )
+
+        # Do the local solve
+
+        solve_out = nosource_local_solve_stage_uniform_2D_ItI(
+            pde_problem=pde_problem
+        )
+
+        Y, R, Phi = solve_out
+
+        # Compute the particular solution
+        n_cheby_bdry = 4 * pde_problem.domain.p - 4
+        source = test_case[K_SOURCE](pde_problem.domain.interior_points)[
+            :, n_cheby_bdry:
+        ]
+        logging.debug(
[…]
