### 2025-07-30 4ad98e3
Removed incoming part impedance computation and fixed accuracy tests.
