+++ tests/test_accuracy/test_ItI_up_pass_accuracy.py
@@ -2,8 +2,6 @@ import logging
 from typing import Dict
 import jax.numpy as jnp
 import jax
-import matplotlib.pyplot as plt
-import pytest
 
 from jaxhps._discretization_tree import DiscretizationNode2D
 from jaxhps._domain import Domain
@@ -21,14 +19,14 @@ from .cases import (
     YMIN,
     YMAX,
     ETA,
-    TEST_CASE_HELMHOLTZ_ITI_COMPLEX_COEFFS,
+    TEST_CASE_ITI_PART_HOMOG,
     K_XX_COEFF,
     K_YY_COEFF,
     K_SOURCE,
-    K_SOLN,
-    K_DUDX,
-    K_DUDY,
     K_I_COEFF,
+    K_PART_SOLN_DUDY,
+    K_PART_SOLN_DUDX,
+    K_PART_SOLN,
 )
 
 ATOL_NONPOLY = 1e-8
@@ -75,51 +73,46 @@ def check_up_pass_accuracy_2D_ItI_uniform_Helmholtz_like(
     # Do an up pass to compute the incoming and outgoing impedance data
     source = test_case[K_SOURCE](domain.interior_points)
 
-    _, _, h_last, g_last = up_pass_uniform_2D_ItI(
+    _, _, h_last = up_pass_uniform_2D_ItI(
         source=source, pde_problem=pde_problem, return_bdry_imp_data=True
     )
 
     # Assemble incoming impedance data
     q = domain.boundary_points.shape[0] // 4
-    boundary_u = test_case[K_SOLN](domain.boundary_points)
+    boundary_u = test_case[K_PART_SOLN](domain.boundary_points)
     boundary_u_normals = jnp.concatenate(
         [
-            -1 * test_case[K_DUDY](domain.boundary_points[:q]),
-            test_case[K_DUDX](domain.boundary_points[q : 2 * q]),
-            test_case[K_DUDY](domain.boundary_points[2 * q : 3 * q]),
-            -1 * test_case[K_DUDX](domain.boundary_points[3 * q :]),
+            -1 * test_case[K_PART_SOLN_DUDY](domain.boundary_points[:q]),
+            test_case[K_PART_SOLN_DUDX](domain.boundary_points[q : 2 * q]),
+            test_case[K_PART_SOLN_DUDY](domain.boundary_points[2 * q : 3 * q]),
+            -1 * test_case[K_PART_SOLN_DUDX](domain.boundary_points[3 * q :]),
         ]
     )
-    incoming_imp_data = boundary_u_normals + 1j * pde_problem.eta * boundary_u
     outgoing_imp_data = boundary_u_normals - 1j * pde_problem.eta * boundary_u
 
-    # Plot outgoing_imp_data against h_last
-    plt.plot(outgoing_imp_data.real, label="Outgoing Impedance Data.real")
-    plt.plot(outgoing_imp_data.imag, label="Outgoing Impedance Data.imag")
-    # plt.plot(incoming_imp_data.real, label="Incoming Impedance Data.real")
-    # plt.plot(incoming_imp_data.imag, label="Incoming Impedance Data.imag")
-    plt.plot(h_last.real, label="h_last.real")
-    plt.plot(h_last.imag, label="h_last.imag")
-    # plt.plot(g_last.real, label="g_last.real")
-    # plt.plot(g_last.imag, label="g_last.imag")
-    plt.legend()
-    plt.show()
+    # Plot the magnitudes
+    # plt.plot(jnp.abs(h_last), ".-", label="computed out")
+    # plt.plot(jnp.abs(outgoing_imp_data), "x-", label="expected out")
+    # # plt.plot(jnp.abs(incoming_imp_data), label="expected in")
+    # plt.legend()
+    # plt.show()
 
     # Check the outgoing impedance data
[…]
