### 2025-08-11 #8549 @dc-luo
Hi @meliao @AnjaliSandip,

I've had some time to play around with the library. This seems to be a great addition to both the PDEs and scientific machine learning communities, supporting GPU-based differentiable solvers for PDEs, especially for challenging problems such as high-frequency Helmholtz. The code is well documented and installation instructions were easy to follow.

I did run into the following issues while installing and running the examples

*Examples*: I've submitted a github issue for these
- [ ] The MATLAB script `driver_gen_SD_matrices.m` for generating the discretization [saves to](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/driver_gen_SD_matrices.m#L17) `../data/examples/SD_matrices`. This is assumed to exist but does not when the repo freshly cloned. Can a call to `mkdir('../data/examples/SD_matrices')` be added to create this directory upon running, just to make it simpler for new users to get started?

- [ ] In the inverse scattering example, the [default](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/inverse_scattering_utils.py#L24) `k` for loading the SD matrices is `k = 20`, which seems to differ from the [default](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/driver_gen_SD_matrices.m#L6) in `driver_gen_SD_matrices.m` (`k = 100`). Is it possible to add this case to be generated by `driver_gen_SD_matrices.m` or to note this in the instructions for running this example?

I notice the path setup for the inverse scattering example is incorrect in the `JOSS_submission` branch but has been fixed in the main branch so I've referred to that here instead.

*Installation*:
- [ ] By default, doing `pip install jaxhps` seems to install a CPU-only version of Jax if it is not already installed. I suggest updating the installation instructions to explicitly write out the main dependencies, and mention how this can be set up with GPU enabled.

I've left the examples and installation checkboxes open in the review checklist and I'll close them out once these two points are addressed.

---

I also had the following thoughts for the examples as a new user experimenting with the package. These are secondary and are more suggestions to the authors (I'd be happy to complete the review without these)
- The use case with the simple poisson problem is great for getting started. However, I did have some initial confusion as to how to best plot the solution at the end (e.g., how is the solution array ordered, does it contain the boundary points, etc.). I think it would be helpful if there was a quick snippet of plotting code at the end of the of that section.
- Many of the examples in the main `examples` directory rely on the MATLAB code for generating the required matrices. While this is not a fundamental issue, I'm wondering if it's possible to pre-assemble these matrices and make them available within the example so users don't have to setup MATLAB (seeing as this is primarily a python package), perhaps just for some of the coarser discretizations with smaller file sizes.
- When running 3D adaptive examples, I was running out of memory on an A100 40GB. Is this expected? While I'd expect the 3D direct solver to be more memory intensive, it might be worth noting down in the documentation as to how much I'd need to run the example.
