### 2025-08-11 meliao/jaxhps#12 @dc-luo
File loading issues while running examples
I ran into the following issues with writing/loading the `SD_matrices` while running the `examples` following the documentation:

- The MATLAB script `driver_gen_SD_matrices.m` for generating the discretization [saves to](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/driver_gen_SD_matrices.m#L17) `../data/examples/SD_matrices`. This is assumed to exist but does not when the repo freshly cloned. Can a call to `mkdir('../data/examples/SD_matrices')` be added to create this directory upon running, just to make it simpler for new users to get started?

- In the inverse scattering example, the [default](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/inverse_scattering_utils.py#L24) `k` for loading the SD matrices is `k = 20`, which seems to differ from the [default](https://github.com/meliao/jaxhps/blob/3e002d052b065c46978fb0aa25a0da6d681532ea/examples/driver_gen_SD_matrices.m#L6) in `driver_gen_SD_matrices.m` (`k = 100`). Is it possible to add this case to be generated by `driver_gen_SD_matrices.m` or to note this in the instructions for running this example?
