### 2025-08-14 f72196a
Update README and docs with installation instructions for GPU systems
