### 2025-08-14 461ec92
Update usage example to include plotting solution
