### 2025-08-14 f678fbc
Changes to make the adaptive 3D experiments run on a 40GB GPU
