### 2025-08-14 312817f
GPU chunksize logic not evaluated on a CPU system
