### 2025-08-14 meliao/ha-hps#12 @meliao
Hi @dc-luo , thank you for your careful review of the repository. I introduced a pull request #13  which answers these points, as well as the points you made in the JOSS review thread. Here are the updates in this PR:

 * By default, the MATLAB script now generates S and D matrices for k=20 and k=100, which are the two wavenumbers used in the inverse and forward scattering examples, respectively.
 * The MATLAB script creates a directory at data/examples/SD_matrices/ if one does not already exist.
 * Changed installation instructions to include instructions for a GPU-compatible version of jax.
 * Extended the "usage quickstart" example to include interpolating the solution to a regular grid and plotting.
 * Fixed the 3D adaptive solve examples, so they can now run on a 40GB GPU.

One point is outstanding from your review:

> Many of the examples in the main examples directory rely on the MATLAB code for generating the required matrices. While this is not a fundamental issue, I'm wondering if it's possible to pre-assemble these matrices and make them available within the example so users don't have to setup MATLAB (seeing as this is primarily a python package), perhaps just for some of the coarser discretizations with smaller file sizes.

I want to ask your advice... what do you think is the best way to share these matrices? Upload them to GitHub as part of the source repository? Google Drive? The ones corresponding to coarser discretizations (`L<=3`) take up about 50MB in total, and no individual file is larger than 10MB. This file size is allowed on GitHub.
