### 2025-10-03 meliao/ha-hps#12 @dc-luo
@meliao Sorry for the delay in responding to this!

Changes to the documentation and source code are very helpful. I have now been able to run the adaptive solves on smaller GPUs.

As for storing the files, I don't have a definitive answer (I've seen projects with both).
- I think if the data files are small and aren't going to change with updates very often, then you could check them into the repository (This could add a little download time). However, if you anticipate these may change, then they could generate large diffs that you might want to avoid.
- If you have a google drive or cloud storage that's _relatively_ permanent then that's also an option (helps to keep things separate). As the other reviewer suggested, zenodo might also be a way of doing this.
