### 2025-10-03 #8549 @dc-luo
> [@dc-luo](https://github.com/dc-luo), [@mrava87](https://github.com/mrava87), [@abhijit-c](https://github.com/abhijit-c), is the paper/software ready for publication?

@abhijit-c Yes, I think the paper/repo is good to publish.

@meliao Sorry for the delay. I have just moved institutions and recently looped back to the [issue](https://github.com/meliao/jaxhps/issues/12). I tested the new code on a 48GB GPU and was able to run the adaptive examples. The documentation looks great too.
