### 2025-10-03 b56868e
Updated docs to reference Zenodo archive storing S and D matrices. Also added GPU VRAM requirements to the docs
