### 2025-10-03 meliao/ha-hps#12 @meliao
Thanks @dc-luo! I uploaded the data files to Zenodo here: [https://doi.org/10.5281/zenodo.17259087](https://doi.org/10.5281/zenodo.17259087). I am going to close this issue and merge the pull request.
