### 2025-10-30 #8549 @meliao
@AnjaliSandip, my coauthors and I have been running some experiments using automatic differentiation applied to our solver in more general settings than the [autodiff example](https://github.com/meliao/jaxhps/blob/main/examples/inverse_wave_scattering.py) currently in the repository. The experiments trying to characterize the accuracy of autodiff in general settings have been inconclusive so far, and we would like to make sure everything is accurate before publishing the paper. I can give you another update at the end of next week, if that is acceptable.
