+++ examples/check_autodiff_vJp.py
@@ -1,0 +1,486 @@
+import argparse
+import os
+import logging
+from functools import partial
+
+import jax.numpy as jnp
+import jax
+import numpy as np
+
+import matplotlib.pyplot as plt
+
+from jaxhps import (
+    DiscretizationNode2D,
+    Domain,
+    PDEProblem,
+)
+from scipy import special
+
+from jaxhps.down_pass import down_pass_uniform_2D_ItI
+from jaxhps.quadrature import chebyshev_weights
+from scattering_potentials import q_GBM_1
+from wave_scattering_utils import (
+    load_SD_matrices,
+    get_uin,
+)
+from sine_transform import (
+    nu_sinetransform,
+    adjoint_nu_sinetransform,
+    get_freqs_up_to_2k,
+)
+from gen_SD_exterior import get_ring_points
+
+from check_autodiff_Jvp import (
+    XMIN,
+    XMAX,
+    YMIN,
+    YMAX,
+    coeffs_to_uscat,
+    eval_uscat_with_source_on_bdry,
+)
+
+plt.rc("font", **{"family": "serif", "serif": ["Computer Modern"]})
+plt.rc("text", usetex=True)
+
+
+def setup_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Compute reference solution for wave scattering problem."
+    )
+
+    parser.add_argument(
+        "-n_pixels",
+        type=int,
+        default=256,
+        help="Number of pixels in the regular grid for the potential.",
+    )
+    parser.add_argument(
+        "-l_vals",
+        type=int,
+        nargs="+",
+        default=[1, 2, 3, 4, 5],
+        help="Number of levels in the quadtree.",
+    )
+    parser.add_argument(
+        "-p",
+        type=int,
+        default=16,
+        help="Chebyshev polynomial order.",
+    )
+
+    parser.add_argument("-k", type=float, default=20.0, help="Wavenumber.")
+    parser.add_argument(
+        "--debug", action="store_true", help="Enable debug logging."
+    )
+    parser.add_argument(
+        "--SD_matrix_prefix",
+        default="../data/examples/SD_matrices",
+    )
+    return parser.parse_args()
+
+
+def hankel_evals(
+    k: float, rec_pts: jax.Array, eval_pts: jax.Array
+) -> jax.Array:
+    """
+    Computes - H_0^2(k|x-y|) for all x in rec_pts and y in eval_pts.
+
+    Args:
+        k (float): Wavenumber
+        rec_pts (jax.Array): Has shape (nrec, 2)
+        eval_pts (jax.Array): Has shape (n_leaves, p**2, 2)
+
+    Returns:
+        jax.Array: Has shape (n_leaves, p**2, nrec)
+    """
+    # Compute the distances between each eval point and each receiver point
+    # Has shape (n_leaves, p**2, nrec, 2)
+    diffs = rec_pts[None, None, ...] - eval_pts[:, :, None]
+    # Compute the Euclidean distance
+    # Has shape (n_leaves, p**2, nrec)
+    distances = jnp.linalg.norm(diffs, axis=-1)
+    # Compute the Hankel function of the second kind
+    # Has shape (n_leaves, p**2, nrec)
+    hankel_evals = -1 * special.hankel2(0, k * distances)
+    return hankel_evals
+
+
+def compute_phi_1(
+    problem: PDEProblem,
+    duscat: jax.Array,
+    nrec: int = 100,
+    rad: float = 5.0,
+) -> jax.Array:
+    """
+
[…]
