+++ examples/check_autodiff_Jvp.py
@@ -1,0 +1,488 @@
+import argparse
+import os
+import logging
+
+import jax.numpy as jnp
+import jax
+import numpy as np
+
+import matplotlib.pyplot as plt
+
+from jaxhps import (
+    DiscretizationNode2D,
+    Domain,
+    PDEProblem,
+    build_solver,
+    solve,
+)
+from jaxhps.up_pass import up_pass_uniform_2D_ItI
+from scattering_potentials import q_GBM_1
+from wave_scattering_utils import (
+    setup_scattering_lin_system,
+    get_uin_and_normals,
+    load_SD_matrices,
+    get_DtN_from_ItI,
+    get_uin,
+)
+from sine_transform import (
+    nu_sinetransform,
+    adjoint_nu_sinetransform,
+    get_freqs_up_to_2k,
+)
+from gen_SD_exterior import gen_S_exterior, gen_D_exterior
+
+XMIN = -1
+XMAX = 1
+YMIN = -1
+YMAX = 1
+
+plt.rc("font", **{"family": "serif", "serif": ["Computer Modern"]})
+plt.rc("text", usetex=True)
+
+
+def setup_args() -> argparse.Namespace:
+    parser = argparse.ArgumentParser(
+        description="Compute reference solution for wave scattering problem."
+    )
+
+    parser.add_argument(
+        "-l_vals",
+        type=int,
+        nargs="+",
+        default=[1, 2, 3, 4, 5],
+        help="Number of levels in the quadtree.",
+    )
+    parser.add_argument(
+        "-p",
+        type=int,
+        default=16,
+        help="Chebyshev polynomial order.",
+    )
+    parser.add_argument(
+        "-n_pixels",
+        type=int,
+        default=500,
+        help="Number of pixels in each dimension to compare solutions.",
+    )
+    parser.add_argument("-k", type=float, default=20.0, help="Wavenumber.")
+    parser.add_argument(
+        "--debug", action="store_true", help="Enable debug logging."
+    )
+    parser.add_argument(
+        "--SD_matrix_prefix",
+        default="../data/examples/SD_matrices",
+    )
+    return parser.parse_args()
+
+
+def coeffs_to_uscat(
+    coeffs: jax.Array,
+    freqs: jax.Array,
+    problem: PDEProblem,
+    source_dirs: jax.Array,
+    S_int: jax.Array,
+    D_int: jax.Array,
+    nrec: int = 100,
+    rad: float = 5.0,
+    return_DtN: bool = False,
+    eval_on_interior: bool = False,
+) -> jax.Array:
+    # First, go from global sine series coefficients to
+    # values of the potential on the interior points
+    spatial_points = problem.domain.interior_points
+    q_hps_grid = adjoint_nu_sinetransform(coeffs, freqs, spatial_points)
+
+    # Next, update the problem's coefficients
+    k = problem.eta
+    lap_coeffs = jnp.ones_like(problem.domain.interior_points[..., 0])
+    i_coeffs = (k**2) * (1 + q_hps_grid)
+    problem.update_coefficients(
+        D_xx_coefficients=lap_coeffs,
+        D_yy_coefficients=lap_coeffs,
+        I_coefficients=i_coeffs,
+    )
+    zero_source = jnp.zeros_like(problem.domain.interior_points[..., 0])
+
+    # Build the solver
+    T_ItI = build_solver(
+        pde_problem=problem,
+        host_device=jax.devices()[0],
+        compute_device=jax.devices()[0],
+        return_top_T=True,
+    )
+
[…]

[…]
