### 2025-11-03 6fd090a
Ready to experiment with inverse problem example on GPU
