+++ examples/inverse_wave_scattering.py
@@ -9,36 +9,52 @@ import numpy as np
 from scipy.sparse.linalg import LinearOperator, lsqr
 from scipy.io import savemat
 
+from jaxhps import DiscretizationNode2D, Domain, PDEProblem
 
-from inverse_scattering_utils import (
-    q_point_sources,
-    source_locations_to_scattered_field,
-    forward_model,
-    SAMPLE_DOMAIN,
-    K,
-    XMIN,
-    XMAX,
-    YMIN,
-    YMAX,
-    OBSERVATION_BOOLS,
-    SOURCE_DIRS,
+# from inverse_scattering_utils import (
+#     SAMPLE_DOMAIN,
+#     K,
+#     XMIN,
+#     XMAX,
+#     YMIN,
+#     YMAX,
+#     OBSERVATION_BOOLS,
+#     SOURCE_DIRS,
+# )
+from check_autodiff_Jvp import coeffs_to_uscat
+from check_autodiff_vJp import quad_weights_hps_panel
+from wave_scattering_utils import load_SD_matrices
+from plotting_utils import (
+    make_scaled_colorbar,
+    TICKSIZE,
+    FONTSIZE,
+    FIGSIZE,
+    get_discrete_cmap,
+    parula_cmap,
 )
-from wave_scattering_utils import get_uin
-from plotting_utils import make_scaled_colorbar, TICKSIZE, FONTSIZE, FIGSIZE
+from scattering_potentials import q_gaussian_bumps
+from sine_transform import (
+    get_freqs_up_to_2k,
+    nu_sinetransform,
+)
 
-# Disable all matplorlib logging
-# logging.getLogger("matplotlib").setLevel(logging.WARNING)
-# logging.getLogger("PIL").setLevel(logging.WARNING)
+# Disable all matplotlib logging
 logging.getLogger("matplotlib").disabled = True
 logging.getLogger("matplotlib.font_manager").disabled = True
 
 jax.config.update("jax_default_device", jax.devices("cpu")[0])
 
+plt.rc("font", **{"family": "serif", "serif": ["Computer Modern"]})
+plt.rc("text", usetex=True)
 
 # Uncomment for debugging NaNs. Slows code down.
 # jax.config.update("jax_debug_nans", True)
 
-N_BUMPS = 4
+XMIN = -1
+XMAX = 1
+YMIN = -1
+YMAX = 1
+SOURCE_DIRS = jnp.array([0.0])
 
 
 def setup_args() -> argparse.Namespace:
@@ -50,8 +66,13 @@ def setup_args() -> argparse.Namespace:
         default="data/examples/inverse_wave_scattering",
     )
     parser.add_argument("--debug", action="store_true")
-    parser.add_argument("--n_iter", type=int, default=10)
-    parser.add_argument("--n_per_side", type=int, default=10)
+    parser.add_argument("--n_iter", type=int, default=20)
+    parser.add_argument("-L", type=int, default=3)
+    parser.add_argument("-p", type=int, default=16)
+    parser.add_argument("-k", type=float, default=20.0)
+    parser.add_argument(
+        "-SD_matrix_fp", type=str, default="../data/examples/SD_matrices"
+    )
 
     return parser.parse_args()
 
@@ -82,85 +103,47 @@ def plot_uscat(
     plt.clf()
 
 
-def plot_iterates(
-    iterates: jnp.array, plot_fp: str, q_evals: jnp.array
+def plot_coeffs(
+    coeffs_ground_truth: jax.Array,
+    coeffs_estimate: jax.Array,
+    freqs: jax.Array,
+    plot_fp: str,
 ) -> None:
     """
-    Make a plot of the evaluations of the ground-truth q and then draw arrows showing the iterates.
+    Plots the absolute value of the components of the coefficient vector against
[…]
