### 2025-11-03 560426f
Finished autodiff experiment for paper
