### 2025-11-06 0e7b557
Updating affiliations for JOSS paper
