### 2025-11-12 #8549 @meliao
@AnjaliSandip I apologize for the delayed update! We have a new example that shows a setting when forward and reverse automatic differentiation converge at high order to the continuous Fréchet derivative when increasing the number of discretization points. This experiment currently lives in a [new branch](https://github.com/meliao/jaxhps/tree/update_autodiff_example); I'd like to update the documentation and pull this branch into main before tagging a release version. I plan to have this finished by the end of the week.
