### 2025-11-13 75f60fe
Updated documentation to reflect new autodiff experiments
