### 2025-11-14 2ff9c55
Bump version to 0.2
