+++ src/jaxhps/__init__.py
@@ -32,4 +32,4 @@ __all__ = [
     "upward_pass_subtree",
     "downward_pass_subtree",
 ]
-__version__ = "0.1"
+__version__ = "0.2"
