### 2025-11-14 #8549 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
